Move titlepanel into subdir
