+++ app/src/components/react-resizable-panels/TitledPanel.tsx
