+++ app/src/components/react-resizable-panels/index.ts
@@ -1,0 +1,1 @@
+export * from "./TitledPanel";
